pyproject.toml
[tool.poetry]
name = "blobby"
version = "0.1.0"
description = "based on https://www.npmjs.com/package/openai-chat/v/1.2.2"
authors = ["devops117 <55235206+devops117@users.noreply.github.com>"]
license = "GPLv3"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
openai = "^0.25.0"
python-telegram-bot = {git = "https://github.com/python-telegram-bot/python-telegram-bot.git", rev = "v20.0a4"}
uvloop = {version = "^0.17.0", optional = true}
cachetools = "^5.2.0"
attrs = "^22.1.0"

[tool.poetry.extras]
speedups = ["uvloop"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
